==> 01b_feature_user.ipynb <==
import pandas as pd
import numpy as np

### data

dtypes = {}
dtypes['ip'] = 'uint32'
dtypes['app'] = 'uint16'
dtypes['device'] = 'uint16'
dtypes['os'] = 'uint16'
dtypes['channel'] = 'uint16'
usecols = list(dtypes.keys()) 
train_data = pd.read_csv('../data/download/train.csv', dtype=dtypes, usecols=usecols)
train_data['click_id'] = train_data.index * -1
train_data = train_data[['click_id','ip','app','os','device','channel']]

dtypes = {}
dtypes['ip'] = 'uint32'
dtypes['app'] = 'uint16'
dtypes['device'] = 'uint16'
dtypes['os'] = 'uint16'
dtypes['channel'] = 'uint16'
dtypes['click_id'] = 'uint32'
usecols = list(dtypes.keys())
test_data = pd.read_csv('../data/download/test_supplement.csv', dtype=dtypes, usecols=usecols)
test_data = test_data[['click_id','ip','app','os','device','channel']]

mapping = pd.read_feather('../data/data/files/mapping.feather')

data = train_data.append(test_data)
del train_data, test_data

### user features

user_count = data.groupby(['ip','device','os'])['channel'].count().reset_index()
user_count.columns = ['ip','device','os','user_count']
user_count.to_feather('../data/data/features/user_count.feather')

user_count = data.groupby(['ip','device','os','app'])['channel'].count().reset_index()
user_count.columns = ['ip','device','os','app','user_app_count']
user_count.to_feather('../data/data/features/user_app_count.feather')

### running counts

run_count = data[['click_id','ip','device','os']].copy()
run_count['run_count'] = data.groupby(['ip','device','os']).cumcount()
run_count_train = run_count[run_count['click_id'] < 0]
run_count_test = run_count[run_count['click_id'] >= 0].iloc[1:,:]
run_count_test = run_count_test.rename(columns={'click_id':'old_id'})
run_count_test = run_count_test.merge(mapping, on='old_id').drop('old_id', axis=1)
run_count_train = run_count_train[['click_id', 'run_count']]
run_count_train['click_id'] = -1 * run_count_train['click_id']
run_count_train['click_id'] = run_count_train['click_id'].astype('uint32')
run_count_test = run_count_test[['click_id', 'run_count']]
run_count_train = run_count_train.reset_index(drop=True)
run_count_test = run_count_test.reset_index(drop=True)
run_count_train.to_feather('../data/data/features/run_count_train.feather')
run_count_test.to_feather('../data/data/features/run_count_test.feather')

del run_count, run_count_train, run_count_test

run_count = data[['click_id','ip','device','os','app']].copy()
run_count['run_app_count'] = data.groupby(['ip','device','os','app']).cumcount()

run_count_train = run_count[run_count['click_id'] < 0]
run_count_test = run_count[run_count['click_id'] >= 0].iloc[1:,:]
run_count_test = run_count_test.rename(columns={'click_id':'old_id'})
run_count_test = run_count_test.merge(mapping, on='old_id').drop('old_id', axis=1)
run_count_train = run_count_train[['click_id', 'run_app_count']]
run_count_train['click_id'] = -1 * run_count_train['click_id']
run_count_train['click_id'] = run_count_train['click_id'].astype('uint32')
run_count_test = run_count_test[['click_id', 'run_app_count']]
run_count_train = run_count_train.reset_index(drop=True)
run_count_test = run_count_test.reset_index(drop=True)

run_count_train.to_feather('../data/data/features/run_app_count_train.feather')
run_count_test.to_feather('../data/data/features/run_app_count_test.feather')
