==> 02a_merge_data.ipynb <==
import pandas as pd
import numpy as np

# one way counts
ip = pd.read_feather('../data/data/features/ip.feather')
os = pd.read_feather('../data/data/features/os.feather')
device = pd.read_feather('../data/data/features/device.feather')
app = pd.read_feather('../data/data/features/app.feather')
channel = pd.read_feather('../data/data/features/channel.feather')
# n-way counts
ip_day_hour = pd.read_feather('../data/data/features/ip_day_hour.feather')
ip_device = pd.read_feather('../data/data/features/ip_device.feather')
ip_app = pd.read_feather('../data/data/features/ip_app.feather')
ip_app_os = pd.read_feather('../data/data/features/ip_app_os.feather')
app_channel = pd.read_feather('../data/data/features/app_channel.feather')
user = pd.read_feather('../data/data/features/user_count.feather')
user_app = pd.read_feather('../data/data/features/user_app_count.feather')
# running counts
run_train = pd.read_feather('../data/data/features/run_count_train.feather')
run_app_train = pd.read_feather('../data/data/features/run_app_count_train.feather')
run_test = pd.read_feather('../data/data/features/run_count_test.feather')
run_app_test = pd.read_feather('../data/data/features/run_app_count_test.feather')
next_click_train = pd.read_feather('../data/data/features/next_click_train.feather')
next_click_test = pd.read_feather('../data/data/features/next_click_test.feather')

def merge(data, mode='train'):
    print('merge starts..', data.shape)
    # one way counts
    data = data.merge(ip, on='ip')
    data = data.merge(os, on='os')
    data = data.merge(app, on='app')
    data = data.merge(channel, on='channel')
    data = data.merge(device, on='device')
    print('one way merge complete...', data.shape)
    # n way counts
    data = data.merge(ip_day_hour, on=['ip','day','hour'])
    data = data.merge(ip_device, on=['ip','device'])
    data = data.merge(ip_app, on=['ip','app'])
    data = data.merge(ip_app_os, on=['ip','app','os'])
    data = data.merge(app_channel, on=['app','channel'])
    print('n way merge complete...', data.shape)
    # user counts
    data = data.merge(user, on=['ip','device','os'])
    data = data.merge(user_app, on=['ip','device','os','app'])
    print('user counts merge complete...', data.shape)
    # run counts
    if mode == 'train':
        data = data.merge(run_train, on='click_id')
        data = data.merge(run_app_train, on='click_id')
        data = data.merge(next_click_train, on='click_id')
        print('id merge complete...', data.shape)
    else:
        data = data.merge(run_test, on='click_id', how='left')
        data = data.merge(run_app_test, on='click_id', how='left')
        data = data.merge(next_click_test, on='click_id', how='left')
        print('id merge complete...', data.shape)
    data = data.reset_index(drop=True)
    return data

train_data = pd.read_feather('../data/data/files/train_data.feather')
train_data = merge(train_data)
train_data.to_feather('../data/data/models/train_data.feather')
del train_data

valid_data = pd.read_feather('../data/data/files/valid_data.feather')
valid_data = merge(valid_data)
valid_data.to_feather('../data/data/models/valid_data.feather')
del valid_data

score_data = pd.read_feather('../data/data/files/score_data.feather')
score_data = merge(score_data,'other')
score_data.to_feather('../data/data/models/score_data.feather')
del score_data
