==> 03a_model.ipynb <==
import pandas as pd
import numpy as np
import lightgbm as lgb

categorical = ['ip','app','channel','device','os']

train_data = pd.read_feather('../data/data/models/train_data.feather', nthreads=15)
train_labels = train_data['is_attributed'].values
train_data = train_data.drop(['is_attributed','day','click_id'], axis=1)

valid_data = pd.read_feather('../data/data/models/valid_data.feather')
valid_labels = valid_data['is_attributed'].values
valid_data = valid_data.drop(['is_attributed','day','click_id'], axis=1)

train_matrix = lgb.Dataset(train_data.values, label=train_labels,feature_name=list(train_data.columns),
                        categorical_feature=categorical)
valid_matrix = lgb.Dataset(valid_data.values, label=valid_labels,feature_name=list(valid_data.columns),
                        categorical_feature=categorical)

booster = {}
booster['boosting_type'] = 'gbdt'
booster['objective'] = 'binary'
booster['learning_rate'] = 0.1
booster['num_leaves'] = 64
booster['max_depth'] = 5
booster['min_child_samples'] = 100
booster['max_bin'] = 100
booster['subsample'] = 0.7
booster['subsample_freq'] = 1
booster['colsample_bytree'] = 0.7
booster['min_child_weight'] = 0
booster['min_split_gain'] = 0
booster['nthread'] = 15
booster['verbose'] = 0
booster['scale_pos_weight'] = 99.7
booster['metric'] = 'auc'

params = {}
params['params'] = booster
params['train_set'] = train_matrix
params['valid_sets'] = [valid_matrix]
params['num_boost_round'] = 1000
params['early_stopping_rounds'] = 50
params['verbose_eval'] = 100

model = lgb.train(**params)

model.save_model('../data/data/models/lightgbm.model')

importance = model.feature_importance(importance_type='split')
importance = pd.DataFrame(importance, columns=['importance'])
importance['feature'] = list(train_data.columns)
importance['importance'] = importance['importance'] / importance['importance'].max()
importance = importance[['feature', 'importance']]
importance = importance.sort_values(by='importance', ascending=False)
importance.head(10)
